% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iUSEiSEE-pkg.R
\docType{package}
\name{iUSEiSEE-pkg}
\alias{iUSEiSEE}
\alias{iUSEiSEE-package}
\alias{iUSEiSEE-pkg}
\title{iUSEiSEE: Learning to use the package iSEE}
\description{
This package provides the material for a fully fledged one-day workshop on the iSEE package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://imbeimainz.github.io/iUSEiSEE/}
  \item Report bugs at \url{https://github.com/imbeimainz/iUSEiSEE/issues}
}

}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})

}
\keyword{internal}
